requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
httpx==0.25.1
pydantic==2.5.0
nats-py==2.0.0
python-dotenv==1.0.0